Refactor to role specific tests
