+++ app/Http/Controllers/Managers/UnretireController.php
@@ -1,0 +1,24 @@
+<?php
+
+namespace App\Http\Controllers\Managers;
+
+use App\Models\Manager;
+use App\Http\Controllers\Controller;
+
+class UnretireController extends Controller
+{
+    /**
+     * Unretire a manager.
+     *
+     * @param  \App\Models\Manager  $manager
+     * @return \lluminate\Http\RedirectResponse
+     */
+    public function __invoke(Manager $manager)
+    {
+        $this->authorize('unretire', $manager);
+
+        $manager->unretire();
+
+        return redirect()->route('managers.index');
+    }
+}
